refactor: convert dict[str, Any] to dataclass types in backend modules
hybrid_magic_model: replace raw dict block construction with Block dataclass.
visual_magic_model_utils: all Block dict params -> Block type, dict access -> dot.
span_block_fix: all Span/Block dict params -> Span/Block type, return Line dataclass.
para_split: fix merge function return types tuple[Block,Block].
pipeline_mkcontent: fix _build_bbox, _inherit_parent_code_render_metadata types.
office_magic_model: add |None to parse_list/parse_index_block.
ocr_utils: widen bbox params to list|tuple for dataclass compat.
span_pre_proc: _cell_range bbox type fix.
pdf_text_tool: remove unused List import.
hybrid_analyze: fix _normalize_page_size, _bbox_to_pixel_bbox return types.

+++ mineru/backend/utils/span_pre_proc.py
@@ -172,7 +172,7 @@ class SpanBlockMatcher:
         self.grid = self._build_grid(self.spans)
 
     @staticmethod
-    def _get_grid_size(spans: list[Span]) -> int:
+    def _get_grid_size(spans: list[Span]) -> float:
         """根据 span 高度估算索引网格大小，避免过细或过粗。"""
         heights = [
             span.bbox[3] - span.bbox[1] for span in spans if span.bbox and span.bbox[3] > span.bbox[1]
